requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCl==1.4.0
pyowm==3.0.0
twilio==6.44.1
youtube-dl==2020.11.12
python-dotenv==0.14.0
async-timeout==3.0.1